requirements.txt
fastapi
uvicorn
requests
pandas
scikit-learn
python-dotenv
joblib